Fixed the dashboard bug
Data now is visible upon npm run dev and refreshing !!! yay
