+++ Xanadu/package.json
@@ -8,8 +8,10 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "chart.js": "^4.4.0",
     "firebase": "^10.4.0",
     "vue": "^3.3.4",
+    "vue-chartjs": "^5.2.0",
     "vue-chartkick": "^1.1.0",
     "vue-router": "^4.2.4",
     "vuetify": "^3.3.20"
