+++ Xanadu/src/components/ActivityChart.vue
@@ -8,57 +8,10 @@
 </template>
 
 <script>
-import firebaseApp from "@/firebase.js";
-import { getFirestore } from "firebase/firestore";
-import { doc, getDocs, collection } from "firebase/firestore";
-const db = getFirestore(firebaseApp);
-
 export default {
   name: "activityChart",
-  data() {
-    return {
-      activityData: {
-        "Water Conservation": 0,
-        "Energy Conservation": 0,
-        "Waste Reduction": 0,
-      },
-      data: { Sun: 32, Mon: 46, Tues: 28 },
-    };
-  },
-  async mounted() {
-    async function getActivityChartData(category) {
-      const subcollectionRefs = [
-        collection(
-          db,
-          "Users/Green Rangers/TestingAcct/Eco-Friendly Activities/" + category
-        ),
-      ];
-      const promises = subcollectionRefs.map((ref) => getDocs(ref));
-      var points = 0;
-      try {
-        const allSnapshots = await Promise.all(promises);
-        allSnapshots.forEach((snapshot) => {
-          snapshot.forEach((doc) => {
-            let document = doc.data();
-            points += Number(document.sustainabilityPoints);
-          });
-        });
-        console.log("Points for " + category + ": ", points);
-      } catch (error) {
-        console.error("Error fetching documents:", error);
-      }
-      return points;
-    }
-    for (const category of [
-      "Water Conservation",
-      "Energy Conservation",
-      "Waste Reduction",
-    ]) {
-      let points = await getActivityChartData(category);
-      console.log(points);
-      this.activityData[category] = points;
-    }
-    console.log(this.activityData);
+  props: {
+    activityData: Object,
   },
 };
 </script>
